fix data_loader

+++ scripts/reazonspeech_manual_downloader.py
@@ -56,9 +56,10 @@ def get_broken_files(target_files):
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(description='Download ReazonSpeech locally.')
     parser.add_argument('-t', '--target', default="tiny", help="tiny/small/medium/large/all", type=str)
-    parser.add_argument('-p', '--pool', default=10, help="thread pool", type=int)
-    parser.add_argument('-s', '--start-que', default=None, help="thread pool", type=int)
-    parser.add_argument('-e', '--end-que', default=None, help="thread pool", type=int)
+    parser.add_argument('-p', '--pool', default=10, type=int)
+    parser.add_argument('-s', '--start-que', default=None, type=int)
+    parser.add_argument('-e', '--end-que', default=None, type=int)
+    parser.add_argument('--health_check', default=None, type=int)
     arg = parser.parse_args()
     target_dir = f"{os.path.expanduser('~')}/.cache/reazon_manual_download/{arg.target}"
     os.makedirs(target_dir, exist_ok=True)
@@ -68,6 +69,14 @@ if __name__ == '__main__':
     if arg.start_que is not None:
         assert arg.end_que is not None
         files = files[arg.start_que:arg.end_que]
+    if arg.health_check:
+        print("check tar files")
+        broken_files = get_broken_files(files)
+        print(f"{len(broken_files)} missing/broken tar file")
+        for i in broken_files:
+            if os.path.exists(i):
+                os.remove(i)
+
     urls = [BASE_URL + DATASET[arg.target]["audio"].format(idx) for idx in files]
     urls.append(BASE_URL + DATASET[arg.target]["tsv"])
     urls = [i for i in urls if not os.path.exists(f"{target_dir}/{arg.target}.{os.path.basename(i)}")]
@@ -91,5 +100,6 @@ if __name__ == '__main__':
         if len(filenames) == 0:
             break
         for i in filenames:
-            os.remove(i)
+            if os.path.exists(i):
+                os.remove(i)
         print(f"retry downloading {len(filenames)} files")
